requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
alembic
pydantic
python-multipart
aiofiles
httpx
python-dotenv
pandas
plotly
jinja2
